pyproject.toml
[tool.poetry]
name = "my-gemini-chatbot"
version = "0.1.0"
packages = [
    { include = "my_gemini_chatbot" }
]
description = ""
authors = ["Rajarshi Karmakar <rkarmaka98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-aiplatform = ">=1.38"
nicegui = "^2.15.0"
jinja2 = "^3.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
